исправила ошибки в редактировании контактной информации
